백단 요청 모두허용상태로
일단 요청은 모두허용

+++ uswine/src/main/java/com/uwa/uswine/user/repository/UserRepository.java
@@ -1,12 +1,12 @@
 package com.uwa.uswine.user.repository;
 
-import com.uwa.uswine.user.entity.UserEntity;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.uwa.uswine.user.entity.UserEntity;
+
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     //username을 받아 DB 테이블에서 회원을 조회하는 메소드
     UserEntity findByEmail(String username);
